fix stuck `signal.wait` by using a deadline instead of a duration
When a SIGUSR1 signal interrupts a `signal.wait` with
`accept_sigusr = false`, the timeout `duration` passed to the recursive
`signal.wait` call doesn't take into account the already elapsed time.

If SIGUSR1 signals are received more frequently than the the timeout
`duration`, the `signal.wait` call will never return because it keeps
getting interrupted and resetting the timeout `duration`.

A simple fix for this issue is to use an absolute `deadline` instead
a relative `duration`. That way, even if the signal gets interrupted,
we'll still trigger a `Timeout` at expected time in subsequent
interations.